refactor api function name
